# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Library API
  version: "1"
servers:
- url: /
security:
- googleIdToken: []
tags:
- name: Me
  description: Utilisateur connecté
- name: Lookup
  description: Recherche de métadonnées
- name: Categories
  description: Catégories de la bibliothèque
- name: Books
  description: "Livres, couvertures et suivi de lecture"
- name: Members
  description: Membres de la bibliothèque
paths:
  /api/categories/{id}:
    put:
      tags:
      - Categories
      operationId: updateCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    delete:
      tags:
      - Categories
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/books/{id}:
    get:
      tags:
      - Books
      operationId: getBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookResponse"
    put:
      tags:
      - Books
      operationId: updateBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookResponse"
    delete:
      tags:
      - Books
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/books/{id}/reading:
    put:
      tags:
      - Books
      operationId: saveMyReading
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReadingResponse"
  /api/books/{id}/cover:
    get:
      tags:
      - Books
      operationId: getCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: size
        in: query
        required: false
        schema:
          type: string
          default: thumb
      - name: key
        in: query
        required: true
        schema:
          type: string
      - name: expires
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: sig
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
    put:
      tags:
      - Books
      operationId: uploadCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoverResponse"
    delete:
      tags:
      - Books
      operationId: deleteCover
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/library/members:
    get:
      tags:
      - Members
      operationId: listMembers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberResponse"
    post:
      tags:
      - Members
      operationId: inviteMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
  /api/categories:
    get:
      tags:
      - Categories
      operationId: listCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags:
      - Categories
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/books:
    get:
      tags:
      - Books
      operationId: listBooks
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: owned
        in: query
        required: false
        schema:
          type: boolean
      - name: category
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: false
        schema:
          $ref: "#/components/schemas/ReadingStatus"
      - name: minRating
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseBookResponse"
    post:
      tags:
      - Books
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookResponse"
  /api/me:
    get:
      tags:
      - Me
      operationId: getMe
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MeResponse"
  /api/lookup/isbn/{isbn}:
    get:
      tags:
      - Lookup
      operationId: lookupIsbn
      parameters:
      - name: isbn
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IsbnLookupResponse"
  /api/library/members/{userId}:
    delete:
      tags:
      - Members
      operationId: removeMember
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
components:
  schemas:
    CategoryRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        color:
          type: string
          pattern: "#[0-9A-Fa-f]{6}"
      required:
      - color
      - name
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
        bookCount:
          type: integer
          format: int64
      required:
      - color
      - id
      - name
    BookRequest:
      type: object
      properties:
        isbn:
          type: string
          maxLength: 20
          minLength: 0
        title:
          type: string
          maxLength: 500
          minLength: 0
        subtitle:
          type: string
          maxLength: 500
          minLength: 0
        authors:
          type: string
          maxLength: 1000
          minLength: 0
        publisher:
          type: string
          maxLength: 300
          minLength: 0
        year:
          type: integer
          format: int32
          maximum: 9999
          minimum: 0
        pages:
          type: integer
          format: int32
          maximum: 100000
          minimum: 1
        language:
          type: string
          maxLength: 10
          minLength: 0
        summary:
          type: string
          maxLength: 20000
          minLength: 0
        owned:
          type: boolean
        categoryIds:
          type: array
          items:
            type: integer
            format: int64
          maxItems: 50
          minItems: 0
      required:
      - title
    BookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        isbn13:
          type: string
        title:
          type: string
        subtitle:
          type: string
        authors:
          type: string
        publisher:
          type: string
        year:
          type: integer
          format: int32
        pages:
          type: integer
          format: int32
        language:
          type: string
        summary:
          type: string
        owned:
          type: boolean
        addedBy:
          type: integer
          format: int64
        addedByName:
          type: string
        createdAt:
          type: string
          format: date-time
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryResponse"
        myReading:
          $ref: "#/components/schemas/ReadingResponse"
        cover:
          $ref: "#/components/schemas/CoverResponse"
        otherReadings:
          type: array
          items:
            $ref: "#/components/schemas/OtherReadingResponse"
      required:
      - categories
      - createdAt
      - id
      - otherReadings
      - owned
      - title
    CoverResponse:
      type: object
      properties:
        thumbUrl:
          type: string
        mediumUrl:
          type: string
      required:
      - mediumUrl
      - thumbUrl
    OtherReadingResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        name:
          type: string
        status:
          $ref: "#/components/schemas/ReadingStatus"
        rating:
          type: integer
          format: int32
        review:
          type: string
      required:
      - name
      - status
      - userId
    ReadingResponse:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReadingStatus"
        rating:
          type: integer
          format: int32
        review:
          type: string
        startedOn:
          type: string
          format: date
        finishedOn:
          type: string
          format: date
      required:
      - status
    ReadingStatus:
      type: string
      enum:
      - TO_READ
      - READING
      - READ
      - ABANDONED
    ReadingRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReadingStatus"
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
        review:
          type: string
          maxLength: 20000
          minLength: 0
        startedOn:
          type: string
          format: date
        finishedOn:
          type: string
          format: date
      required:
      - status
    MemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          minLength: 0
        role:
          $ref: "#/components/schemas/Role"
      required:
      - email
    Role:
      type: string
      enum:
      - OWNER
      - MEMBER
    MemberResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        role:
          $ref: "#/components/schemas/Role"
      required:
      - email
      - role
      - userId
    LibrarySummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
      required:
      - id
      - role
    MeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        libraries:
          type: array
          items:
            $ref: "#/components/schemas/LibrarySummary"
      required:
      - email
      - id
      - libraries
    IsbnLookupResponse:
      type: object
      properties:
        isbn13:
          type: string
        source:
          $ref: "#/components/schemas/MetadataSource"
        existingBookId:
          type: integer
          format: int64
        title:
          type: string
        subtitle:
          type: string
        authors:
          type: string
        publisher:
          type: string
        year:
          type: integer
          format: int32
        pages:
          type: integer
          format: int32
        language:
          type: string
        summary:
          type: string
        coverUrl:
          type: string
      required:
      - isbn13
      - source
      - title
    MetadataSource:
      type: string
      enum:
      - LIBRARY
      - OPEN_LIBRARY
      - GOOGLE_BOOKS
    PageResponseBookResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BookResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
      required:
      - items
      - page
      - size
      - totalItems
      - totalPages
  securitySchemes:
    googleIdToken:
      type: http
      description: ID token Google
      scheme: bearer
      bearerFormat: JWT